feat: make ram check more robust
First reduce RAM size to what is effectively used (1600 bytes is already
over but it is a nice number). Then load data from ram_contents in two
RAMs: one used by the circuit and one 'target'. After, we check for
equality of every single cell of the used and target RAMs.

This method, even if a bit more computationally heavy, it guards against
writes to random addresses that the other check did not cover (such as
(0; 1000), (1255; 3000) and (3255; infty)).